Restore old calculator class.
